Reorganized internal structure
