+++ lib/descope.dart
@@ -2,12 +2,12 @@ library descope_flutter;
 
 import 'package:descope/src/session/session.dart';
 
-import 'src/config.dart';
-import 'src/routes.dart';
-import 'src/sdk.dart';
+import 'src/sdk/config.dart';
+import 'src/sdk/routes.dart';
+import 'src/sdk/sdk.dart';
 import 'src/session/manager.dart';
 
-export 'src/sdk.dart' show DescopeSdk;
+export 'src/sdk/sdk.dart' show DescopeSdk;
 export 'src/session/session.dart' show DescopeSession;
 export 'src/session/token.dart' show DescopeToken;
 export 'src/types/others.dart';
