+++ lib/src/internal/routes/auth.dart
@@ -1,9 +1,8 @@
-import 'package:descope/src/types/responses.dart';
-
+import '/src/sdk/routes.dart';
+import '/src/types/others.dart';
+import '/src/types/responses.dart';
 import '../http/descope_client.dart';
 import '../http/responses.dart';
-import '../routes.dart';
-import '../types/others.dart';
 import 'shared.dart';
 
 class Auth implements DescopeAuth {
