+++ lib/src/internal/routes/sso.dart
@@ -1,7 +1,6 @@
-import 'package:descope/src/types/responses.dart';
-
+import '/src/sdk/routes.dart';
+import '/src/types/responses.dart';
 import '../http/descope_client.dart';
-import '../routes.dart';
 import 'shared.dart';
 
 class Sso implements DescopeSso {
